+++ include/hal_core/netlist/netlist_utils.h
@@ -101,6 +101,60 @@ namespace hal
         CORE_API std::vector<Gate*> get_next_sequential_gates(const Net* net, bool get_successors);
 
         /**
+         * Find all gates on the predecessor or successor path of a gate.
+         * Traverses all input or output nets until gates of the specified base types are found.
+         * The result may include the provided gate itself.
+         * The use of the this cached version is recommended in case of extensive usage to improve performance. 
+         * The cache will be filled by this function and should initially be provided empty.
+         * Different caches for different values of get_successors shall be used.
+         *
+         * @param[in] gate - The initial gate.
+         * @param[in] get_successors - If true, the successor path is returned, otherwise the predecessor path is returned.
+         * @param[in] stop_types - Base types of gate types at which traversal should halt.
+         * @param[inout] cache - The cache. 
+         * @returns All gates on the predecessor or successor path of the gate.
+         */
+        CORE_API std::vector<Gate*> get_path(const Gate* gate, bool get_successors, std::set<GateType::BaseType> stop_types, std::unordered_map<u32, std::vector<Gate*>>& cache);
+
+        /**
+         * Find all gates on the predeccessor or successor path of a gate.
+         * Traverses all input or output nets until gates of the specified base types are found.
+         * The result may include the provided gate itself.
+         *
+         * @param[in] gate - The initial gate.
+         * @param[in] get_successors - If true, the successor path is returned, otherwise the predecessor path is returned.
+         * @param[in] stop_types - Base types of gate types at which traversal should halt.
+         * @returns All gates on the predecessor or successor path of the gate.
+         */
+        CORE_API std::vector<Gate*> get_path(const Gate* gate, bool get_successors, std::set<GateType::BaseType> stop_types);
+
+        /**
+         * Find all gates on the predecessor or successor path of a net.
+         * Traverses all input or output nets until gates of the specified base types are found.
+         * The use of the this cached version is recommended in case of extensive usage to improve performance. 
+         * The cache will be filled by this function and should initially be provided empty.
+         * Different caches for different values of get_successors shall be used.
+         *
+         * @param[in] net - The initial net.
+         * @param[in] get_successors - If true, the successor path is returned, otherwise the predecessor path is returned.
+         * @param[in] stop_types - Base types of gate types at which traversal should halt.
+         * @param[inout] cache - The cache. 
+         * @returns All gates on the predecessor or successor path of the net.
+         */
+        CORE_API std::vector<Gate*> get_path(const Net* net, bool get_successors, std::set<GateType::BaseType> stop_types, std::unordered_map<u32, std::vector<Gate*>>& cache);
+
+        /**
+         * Find all gates on the predecessor or successor path of a net.
+         * Traverses all input or output nets until gates of the specified base types are found.
+         *
+         * @param[in] net - The initial net.
+         * @param[in] get_successors - If true, the successor path is returned, otherwise the predecessor path is returned.
+         * @param[in] stop_types - Base types of gate types at which traversal should halt.
+         * @returns All gates on the predecessor or successor path of the net.
+         */
+        CORE_API std::vector<Gate*> get_path(const Net* net, bool get_successors, std::set<GateType::BaseType> stop_types);
+
+        /**
          * Get the nets that are connected to a subset of pins of the specified gate.
          * 
          * @param[in] gate - The gate.